# workspaces/x2a/plugins/x2a-backend/src/schema/openapi.yaml
openapi: 3.0.4
info:
  title: x2a
  description: The [X2Ansible](https://github.com/redhat-developer/x2ansible) API.
  version: 0.0.1
servers:
  - url: /
paths:
  /projects:
    get:
      summary: Returns a list of projects.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Page number
        - in: query
          name: pageSize
          schema:
            type: integer
          required: false
          description: Page size
        - in: query
          name: order
          schema:
            type: string
            enum:
              - asc
              - desc
          required: false
          description: Sort order, either ascending ("asc") or descending ("desc")
        - in: query
          name: sort
          schema:
            type: string
            enum:
              - createdAt
              - name
              - status
              - description
              - ownedBy
          required: false
          description: Sort by field
      responses:
        '200':
          description: All projects filtered by the user permissions. Sorted and paginated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCount:
                    type: integer
                    description: Total number of projects available after filtering (useful for pagination).
                  items:
                    type: array
                    description: Projects filtered by the user permissions and pagination (if requested).
                    items:
                      $ref: '#/components/schemas/Project'
    post:
      summary: Creates a new project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Full name of the project
                ownedByGroup:
                  type: string
                  description: Optional - group that owns the project instead of the creator (logged in user)
                description:
                  type: string
                  description: Description of the project
                sourceRepoUrl:
                  type: string
                  description: URL of the source repository
                targetRepoUrl:
                  type: string
                  description: URL of the target repository
                sourceRepoBranch:
                  type: string
                  description: Branch of the source repository (default to main)
                targetRepoBranch:
                  type: string
                  description: Branch of the target repository (default to main)
                acceptedRuleIds:
                  type: array
                  items:
                    type: string
                  description: UUIDs of rules the project accepts (required rules auto-appended)
              required:
                - name
                - description
                - sourceRepoUrl
                - targetRepoUrl
                - sourceRepoBranch
                - targetRepoBranch
      responses:
        '201':
          description: Created project data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
  /projects/{projectId}:
    get:
      summary: Returns a project by ID.
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Project data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found.
    delete:
      summary: Deletes a project by ID.
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Project deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deletedCount:
                    type: integer
                    description: Number of projects deleted (0 or 1)
                required:
                  - deletedCount
    patch:
      summary: Updates an existing project.
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  type: string
                  description: Full name of the project
                ownedBy:
                  type: string
                  description: The user who owns the project (Backstage user reference)
                description:
                  type: string
                  description: Description of the project
      responses:
        '200':
          description: Updated project data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid request data.
        '404':
          description: Project not found.
  /projects/{projectId}/run:
    post:
      summary: Triggers the init phase to produce the migration plan
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sourceRepoAuth:
                  $ref: '#/components/schemas/GitRepoAuth'
                targetRepoAuth:
                  $ref: '#/components/schemas/GitRepoAuth'
                aapCredentials:
                  $ref: '#/components/schemas/AAPCredentials'
                userPrompt:
                  type: string
                  description: Optional user prompt for customizing the migration
                refresh:
                  type: boolean
                  description: When true, runs init in refresh mode to resync the module list from an existing migration plan
              required:
                - sourceRepoAuth
                - targetRepoAuth
      responses:
        '200':
          description: Init job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum:
                      - pending
                    description: Job status
                  jobId:
                    type: string
                    description: UUID of the created job
                required:
                  - status
                  - jobId
        '404':
          description: Project not found
...
components:
  securitySchemes:
    callbackSignature:
      type: apiKey
      in: header
      name: X-Callback-Signature
      description: |
        HMAC-SHA256 signature of raw request body using job-specific callback token.
        Signature is computed as: HMAC-SHA256(callbackToken, raw_request_body)
        and sent as a hex-encoded string.

        Replay attack prevention: Requests are only accepted within 3 hours of job creation time.
  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
          description: UUID for the project
        name:
          type: string
          description: Full name of the project
        description:
          type: string
          description: Description of the project
        sourceRepoUrl:
          type: string
          description: URL of the source repository
        targetRepoUrl:
          type: string
          description: URL of the target repository
        sourceRepoBranch:
          description: Branch of the source repository
          type: string
        targetRepoBranch:
          description: Branch of the target repository
          type: string
        createdAt:
          type: string
          format: date-time
          description: Date/time when the project was created
        ownedBy:
          type: string
          description: The user or group who owns the project (Backstage entity reference)
        dirName:
          type: string
          description: Immutable directory name for the project in the target repository (computed once at creation)
        migrationPlan:
          $ref: '#/components/schemas/Artifact'
          description: Project migration plan artifact (by init phase)
        status:
          $ref: '#/components/schemas/ProjectStatus'
          description: Project status calculated from the status of its modules
        initJob:
          $ref: '#/components/schemas/Job'
          description: Project's init job
        acceptedRules:
          type: array
          items:
            $ref: '#/components/schemas/RuleSnapshot'
          description: Snapshot of rules accepted at project creation time
      required:
        - id
        - name
        - sourceRepoUrl
        - targetRepoUrl
        - sourceRepoBranch
        - targetRepoBranch
        - createdAt
        - ownedBy
...
    JobStatusEnum:
      type: string
      enum:
        - pending
        - running
        - success
        - error
        - cancelled
...
    ProjectStatusState:
      type: string
      description: |
        Project status state.
        It is calculated from the status of its modules.
        - created: Project is created but not yet initialized
        - initializing: Project's init job is running or scheduling
        - initialized: Project's init job finished successfully. Either module list is empty or all modules are in pending state.
        - inProgress: At least one module is beyond the pending state.
        - completed: All modules are in success state
        - failed: At least one module is in error state
      enum:
        - created
        - initializing
        - initialized
        - inProgress
        - completed
        - failed
    ModulesStatusSummary:
      type: object
      properties:
        total:
          type: integer
          description: Total number of modules in the project
        finished:
          type: integer
          description: Number of modules in success state of the publish phase (no more work is needed)
        waiting:
          type: integer
          description: Number of modules in success state of a non-publish phase (means waiting for human interaction)
        pending:
          type: integer
          description: Number of modules in pending state (scheduled for execution but not actually running)
        running:
          type: integer
          description: Number of modules in running state (actually running)
        error:
          type: integer
          description: Number of modules in error state (execution is over but failed)
        cancelled:
          type: integer
          description: Number of modules in cancelled state (last job was cancelled by the user)
        removed:
          type: integer
          description: Number of soft-deleted modules (excluded from other counts)
      required:
        - total
        - finished
        - waiting
        - pending
        - running
        - error
        - cancelled
        - removed
    ProjectStatus:
      type: object
      properties:
        state:
          $ref: '#/components/schemas/ProjectStatusState'
        modulesSummary:
          $ref: '#/components/schemas/ModulesStatusSummary'
      required:
        - state
        - modulesSummary
    Job:
      type: object
      required:
        - id
        - projectId
        - startedAt
        - phase
        - k8sJobName
        - status
      properties:
        id:
          type: string
          description: UUID for the job
        projectId:
          type: string
          description: UUID of the owning project
        moduleId:
          type: string
          description: UUID of the owning module (optional)
        startedAt:
          type: string
          format: date-time
          description: Date/time when the job started
        finishedAt:
          type: string
          format: date-time
          description: Date/time when the job finished
        phase:
          $ref: '#/components/schemas/MigrationPhase'
          description: Migration phase of the job
        k8sJobName:
          type: string
          description: Name of the Kubernetes job
        status:
          $ref: '#/components/schemas/JobStatusEnum'
        errorDetails:
          type: string
          description: Detailed error information if the job failed to execute
        artifacts:
          type: array
          description: List of artifacts produced by the job
          items:
            $ref: '#/components/schemas/Artifact'
        telemetry:
          $ref: '#/components/schemas/Telemetry'
        commitId:
          type: string
          description: Git commit SHA produced by this job
        attemptCount:
          type: integer
          description: Number of restarts (attempts) for this projectId+moduleId+phase combination. Calculated value, not persisted in DB.
        firstAttemptAt:
          type: string
          format: date-time
          description: When the very first attempt for this phase was started.
    ArtifactType:
      type: string
      enum:
        - migration_plan
        - module_migration_plan
        - migrated_sources
        - project_metadata
        - ansible_project
    Artifact:
      type: object
      properties:
        id:
          type: string
          description: UUID for the artifact
        type:
          $ref: '#/components/schemas/ArtifactType'
          description: Type of the artifact
        value:
          type: string
          description: Value of the artifact
      required:
        - id
        - type
        - value
    GitRepoAuth:
      type: object
      properties:
        token:
          type: string
          description: Authentication token for the git repository
      required:
        - token
    AAPCredentials:
      type: object
      properties:
        url:
          type: string
          description: Ansible Automation Platform URL
        orgName:
          type: string
          description: AAP organization name
        oauthToken:
          type: string
          description: OAuth token for AAP authentication
        username:
          type: string
          description: Username for AAP authentication (alternative to OAuth token)
        password:
          type: string
          description: Password for AAP authentication (alternative to OAuth token)
      required:
        - url
        - orgName
    MigrationPhase:
      type: string
      enum:
        - init
        - analyze
        - migrate
        - publish
      description: All migration phases
...
    Telemetry:
      type: object
      description: Execution telemetry and metrics for a migration phase
      properties:
        summary:
          type: string
          description: Execution summary from x2a convertor
        phase:
          type: string
          description: Phase name (init, analyze, migrate, publish)
        startedAt:
          type: string
          format: date-time
          description: When the phase started (ISO 8601 format)
        endedAt:
          type: string
          format: date-time
          description: When the phase completed (ISO 8601 format)
        agents:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AgentMetrics'
          description: Per-agent execution metrics (agent name to metrics mapping)
      required:
        - summary
        - phase
        - startedAt
...
    RuleSnapshot:
      type: object
      description: Snapshot of a rule at the time it was accepted by a project
      properties:
        id:
          type: string
          description: UUID of the rule
        title:
          type: string
          description: Title of the rule at acceptance time
        description:
          type: string
          description: Description of the rule at acceptance time
      required:
        - id
        - title
        - description
    AgentMetrics:
      type: object
      description: Telemetry data for a single agent execution within a phase
      properties:
        name:
          type: string
          description: Agent identifier (e.g., "PlanningAgent", "WriteAgent")
        startedAt:
          type: string
          format: date-time
          description: When agent execution started (ISO 8601 format)
        endedAt:
          type: string
          format: date-time
          description: When agent execution completed (ISO 8601 format)
        durationSeconds:
          type: number
          format: double
          description: Execution duration in seconds
        inputTokens:
          type: integer
          description: LLM input token count for this agent
        outputTokens:
          type: integer
          description: LLM output token count for this agent
        metrics:
          type: object
          additionalProperties: true
          description: Custom key-value metrics recorded by the agent
        toolCalls:
          type: object
          additionalProperties:
            type: integer
          description: Tool call counts (tool name to count mapping)
      required:
        - name
        - durationSeconds